feat: implement logging, improve exception handling, and fix Long comparison bug

+++ src/main/java/com/learningmat/ecommerce/dto/request/CartRequest.java
@@ -3,8 +3,6 @@ package com.learningmat.ecommerce.dto.request;
 import jakarta.validation.constraints.Min;
 
 public record CartRequest(
-        @Min(value = 1, message = "PRODUCT_NOT_FOUND")
-        int productId,
-        @Min(value = 1, message = "INVALID_QUANTITY")
-        int quantity) {
+        @Min(value = 1, message = "PRODUCT_NOT_FOUND") Long productId,
+        @Min(value = 1, message = "INVALID_QUANTITY") int quantity) {
 }
